// coordinacion-2-main/src/types.ts
export type SlotType = 'm' | 't' | 'n';
export type DoctorRole = 
  | 'Médico Rural' 
  | 'Médico General' 
  | 'Médico Especialista' 
  | 'Enfermero Jefe' 
  | 'Auxiliar Enfermería' 
  | 'Interno' 
  | 'Triage' 
  | 'Laboratorio' 
  | 'Odontólogo'
  | 'Especialista'
  | 'Fisioterapeuta'
  | 'Rayos X'
  | 'Médico Obstetra/Ginecólogo'
  | 'Jefe de Partos';

export interface Doctor {
  id: number;
  nombre: string;
  apellidos?: string;
  cedula?: string;
  registroMedico?: string;
  email?: string;
  telefono?: string;
  genero?: 'M' | 'F';
  cat: 'Planta' | 'CTA' | 'APS' | 'Rural' | 'Disponibilidad';
  rol: DoctorRole;
  st: 'activo' | 'inactivo';
  contacto?: string;
  username?: string;
  password?: string;
  permissions?: string[];
  passwordLastChanged?: number;
  createdAt?: number;
  mustChangePassword?: boolean;
  sortOrder?: number;
}
// ...
export type ShiftMap = Record<number, string>; // day -> sigla

export interface DoctorShifts {
  m: ShiftMap;
  t: ShiftMap;
  n: ShiftMap;
}

export type MonthlyData = Record<number, DoctorShifts>; // doctorId -> shifts